+++ src/YouAreFriends.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+let YouAreFriends = () => <div className='add-friend'>Friends <i className="far fa-check-circle"></i></div>
+
+export default YouAreFriends;
